Basic HivTest and Cd4Test working together. Only works with one person, cohorts of >2 people appear to fail.

+++ priorityQ/rng.h
@@ -602,7 +602,7 @@ struct Rng
 	inline Doub doub() {return 5.42101086242752217E-20 * int64();}
 	inline Uint int32() {return (Uint)int64();}
 	
-	inline int SampleExpDist(double mean) // Change back to double if this isn't truncating the value as desired.
+	inline double SampleExpDist(double mean)
 	{
 		double exp = 0;
 		double rate;
